Adding file "source\ScriptModules\ScriptModulesNested.h"
Reason, for easier overview.
