+++ AutoHotkeyx.vcxproj
@@ -264,6 +264,7 @@
     <ClInclude Include="source\ScriptModules\ScriptModulesAddNames.h" />
     <ClInclude Include="source\ScriptModules\ScriptModulesDefines.h" />
     <ClInclude Include="source\ScriptModules\ModuleListMethods.h" />
+    <ClInclude Include="source\ScriptModules\ScriptModulesNested.h" />
     <ClInclude Include="source\ScriptModules\ScriptModulesReleaseObjects.h" />
     <ClInclude Include="source\ScriptModules\ScriptModulesSourceFileIndex.h" />
     <ClInclude Include="source\script_com.h" />
